# backend/docs/swagger.yaml
basePath: /api
definitions:
  handlers.CommentDTO:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      text:
        type: string
      user:
        $ref: '#/definitions/handlers.UserPublic'
    type: object
...
  handlers.NotificationDTO:
    properties:
      actor:
        $ref: '#/definitions/handlers.UserPublic'
      createdAt:
        type: string
      id:
        type: integer
      kind:
        type: string
      message:
        type: string
      portfolio:
        properties:
          id:
            type: integer
          title:
            type: string
        type: object
      read:
        type: boolean
    type: object
...
  handlers.RoastDTO:
    properties:
      aiGenerated:
        type: boolean
      body:
        type: string
      createdAt:
        type: string
      helpful:
        type: integer
      id:
        type: integer
      portfolio:
        properties:
          id:
            type: integer
          screenshotUrl:
            type: string
          title:
            type: string
        type: object
      stars:
        type: integer
      title:
        type: string
      user:
        $ref: '#/definitions/handlers.UserPublic'
    type: object
...
  handlers.UserPublic:
    properties:
      bio:
        type: string
      githubUrl:
        type: string
      id:
        type: integer
      initials:
        type: string
      linkedinUrl:
        type: string
      location:
        type: string
      name:
        type: string
      role:
        type: string
      twitterUrl:
        type: string
      username:
        type: string
      websiteUrl:
        type: string
    type: object
...
  handlers.aiGenerateRequest:
    properties:
      mode:
        type: string
      portfolioId:
        type: integer
      prompt:
        type: string
    required:
    - mode
    - portfolioId
    type: object
  handlers.aiGenerateResponse:
    properties:
      text:
        type: string
    type: object
  handlers.aiShareRequest:
    properties:
      mode:
        type: string
      portfolioId:
        type: integer
      text:
        type: string
    required:
    - mode
    - portfolioId
    - text
    type: object
  handlers.authResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/handlers.UserPublic'
    type: object
  handlers.createCommentRequest:
    properties:
      text:
        type: string
    required:
    - text
    type: object
...
  handlers.likeResponse:
    properties:
      likeCount:
        type: integer
      liked:
        type: boolean
    type: object
  handlers.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.registerRequest:
    properties:
      email:
        example: aria@devfolio.dev
        type: string
      name:
        example: Aria Chen
        type: string
      password:
        example: password123
        minLength: 8
        type: string
      username:
        example: ariac
        type: string
    required:
    - email
    - password
    - username
    type: object
...
info:
  contact: {}
  description: REST API powering the DevFolio frontend — auth, portfolios, versions,
    likes, comments, roasts, leaderboard, notifications and AI Studio.
  title: FolioHub API
  version: "1.0"
paths:
  /ai/generate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Target portfolio, mode, optional focus prompt
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.aiGenerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.aiGenerateResponse'
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Generate AI Studio feedback
      tags:
      - ai
  /ai/share:
    post:
      consumes:
      - application/json
      parameters:
      - description: AI result to share
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.aiShareRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.RoastDTO'
      security:
      - BearerAuth: []
      summary: Share an AI Studio result as a roast
      tags:
      - ai
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.authResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserPublic'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: New account details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.authResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an account
      tags:
      - auth
...
  /notifications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.NotificationDTO'
            type: array
      security:
      - BearerAuth: []
      summary: List my notifications
      tags:
      - notifications
  /notifications/read-all:
    post:
      produces:
      - application/json
      responses:
        "204":
          description: no content
      security:
      - BearerAuth: []
      summary: Mark all notifications read
      tags:
      - notifications
...
  /portfolios/{id}/comments:
    get:
      parameters:
      - description: Portfolio ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.CommentDTO'
            type: array
      summary: List a portfolio's comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Portfolio ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment text
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CommentDTO'
      security:
      - BearerAuth: []
      summary: Comment on a portfolio
      tags:
      - comments
  /portfolios/{id}/like:
    post:
      parameters:
      - description: Portfolio ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.likeResponse'
      security:
      - BearerAuth: []
      summary: Like or unlike a portfolio
      tags:
      - portfolios
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"